Safari-609.1.11.1/LayoutTests/fast/dom/HTMLInputElement/null-handling.html

<DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description("Tests null handling of several HTMLInputElement attributes");

var input = document.createElement("input");
shouldBeNull("input.getAttribute('formMethod')", "");
shouldBeNull("input.getAttribute('formEnctype')", "");
evalAndLog("input.formMethod = null");
shouldBeEqualToString("input.getAttribute('formMethod')", "null");
evalAndLog("input.formEnctype = null");
shouldBeEqualToString("input.getAttribute('formEnctype')", "null");
</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
</html>
