LayoutTests/fast/dom/HTMLInputElement/null-handling-expected.txt

Tests null handling of several HTMLInputElement attributes

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.getAttribute('formMethod') is null
PASS input.getAttribute('formEnctype') is null
input.formMethod = null
PASS input.getAttribute('formMethod') is "null"
input.formEnctype = null
PASS input.getAttribute('formEnctype') is "null"
PASS successfullyParsed is true

TEST COMPLETE
